refactor(boundaries): make core an exhaustive layered contract
core was the one app whose internal boundary was enforced by a
hand-maintained `forbidden` contract rather than a clean `layers` one,
because its top-level import graph had a cycle. Break the cycle and
replace the enumerated forbidden with an exhaustive layered contract
mirroring catalog.

Two down-up edges caused the cycle, both model-layer code living above
the models layer:

- `models -> markdown`: `MarkdownField` (a Django field type) lived in
  the markdown rendering package. Move the storage class down into the
  models layer. The wikilink-aware authoring<->storage conversion stays
  in `core/markdown` (it never ran from inside the field), so no runtime
  `models -> markdown`/`wikilinks` edge returns.
- `validators -> models`: `bulk_create_validated` was dead code (zero
  production callers; orphaned when ingest moved to claim_ingest). Delete
  it and its test; `validators` is now pure and model-free.

Split the relocated field package by concern (the two classes share only
the length-CHECK helper):

  core/models/fields/
    _max_length.py  shared char_length CHECK contributor
    text.py         BoundedTextField
    markdown.py     MarkdownField

`apps.core.models` re-exports both classes; `MarkdownField` is no longer
re-exported from the `markdown` package (the one in-tree caller now
imports it from its storage home — no back-compat shim needed in a
monorepo).

Replace `The core api layer is a leaf` with an exhaustive `Core internal
layers` contract. The layer order is derived from the (now acyclic)
graph including indirect reach through other apps. `exhaustive = true`
closes the admin/checks/soft_delete gaps the enumerated source list
silently left; apps/migrations/tests stay exhaustive_ignores as in
catalog. Contract goes green with zero baselines.

Delete docs/plans/CoreInternalLayering.md (the plan this implements) and
update docs/AppBoundaries.md.
