+++ backend/apps/core/markdown/__init__.py
@@ -7,9 +7,9 @@ wikilinks from saved content.
 Layout:
 
 - :mod:`render` — markdown→HTML pipeline + wikilink→markdown-link rendering.
-- :mod:`field` — :class:`MarkdownField` and the conversion path that
-  doesn't touch ``RecordReference``. Models import :class:`MarkdownField`
-  from here (notably the ``DescribedModel`` mixin).
+- :mod:`field` — the wikilink-aware authoring↔storage conversion path
+  that doesn't touch ``RecordReference``. The :class:`MarkdownField`
+  storage class itself lives in :mod:`apps.core.models.fields`.
 - :mod:`references` — the markdown→\\ ``RecordReference`` bridge
   (``sync_references``, ``save_inline_markdown_field``). Save-path
   callers import directly from there; the bridge is intentionally not
@@ -17,7 +17,6 @@ Layout:
 """
 
 from apps.core.markdown.field import (
-    MarkdownField,
     WikilinkAuthoringLookup,
     apply_storage_to_authoring,
     convert_authoring_to_storage,
@@ -37,7 +36,6 @@ from apps.core.markdown.render import (
 )
 
 __all__ = [
-    "MarkdownField",
     "RenderedField",
     "WikilinkAuthoringLookup",
     "apply_storage_to_authoring",
