+++ backend/tests/test_bounded_text_fields.py
@@ -15,8 +15,8 @@ from django.db.models.constraints import CheckConstraint
 
 from apps.accounts.models import User
 from apps.catalog.models import Title
-from apps.core.markdown.field import DEFAULT_MARKDOWN_MAX_LENGTH, MarkdownField
 from apps.core.models import BoundedTextField
+from apps.core.models.fields import DEFAULT_MARKDOWN_MAX_LENGTH, MarkdownField
 from apps.provenance.models import ChangeSet, ChangeSetAction, CitationInstance
 from apps.provenance.models.changeset import CHANGESET_NOTE_MAX_LENGTH
 from apps.provenance.models.citation_instance import (
